+++ backend/src/auth/auth.service.ts
@@ -26,16 +26,6 @@ export class AuthService {
 		return null;
 	}
 
-	async create2fadiv(id: number): Promise<string> {
-		const user: UserEntity | string = await this.userService.getUser(-1, id);
-		if (typeof user == "string") return "";
-		const codedata = twofa.getTwoFactorAuthenticationCode();
-		user.twoFactorSecret = codedata.base32;
-		const qrcode = await twofa.createQrCodeAsURL(codedata.otpauthUrl);
-		await this.userService.saveUser(user);
-		return qrcode;
-	}
-
 	async getQrRetSecret(): Promise<object> {
 		const codedata = twofa.getTwoFactorAuthenticationCode();
 		const qrcode = await twofa.createQrCodeAsURL(codedata.otpauthUrl);
@@ -54,6 +44,16 @@ export class AuthService {
 	async check2faInput(input: string, secret: string): Promise<boolean> {
 		return await twofa.check2faInput(input, secret);
 	}
+
+	// async create2fadiv(id: number): Promise<string> {
+	// 	const user: UserEntity | string = await this.userService.getUser(-1, id);
+	// 	if (typeof user == "string") return "";
+	// 	const codedata = twofa.getTwoFactorAuthenticationCode();
+	// 	user.twoFactorSecret = codedata.base32;
+	// 	const qrcode = await twofa.createQrCodeAsURL(codedata.otpauthUrl);
+	// 	await this.userService.saveUser(user);
+	// 	return qrcode;
+	// }
 
 	// async testProtector(): Promise<void> {
 	// 	const h = await this.protectorService.hash("i like react kidding");
